elimnar amistad

+++ src/amistad/amistad.controller.ts
@@ -1,12 +1,16 @@
 import {
-  Body,
   Controller,
+  Get,
   Post,
   Delete,
-  Get,
+  UseGuards,
+  Req,
   Param,
   ParseIntPipe,
+  Body,
+  BadRequestException,
 } from '@nestjs/common';
+import { AuthGuard } from '@nestjs/passport';
 
 import { AmistadService } from './amistad.service';
 
@@ -27,11 +31,19 @@ export class AmistadController {
     return this.amistadService.createAmistad(id_usuario_a, id_usuario_b);
   }
 
-  @Delete('delete')
+  @UseGuards(AuthGuard('jwt'))
+  @Delete('delete/:id_amigo')
   async deleteAmistad(
-    @Body('id_usuario_a') id_usuario_a: number,
-    @Body('id_usuario_b') id_usuario_b: number,
+    @Req() req,
+    @Param('id_amigo', ParseIntPipe) id_amigo: number,
   ) {
+    const id_usuario_a = req.user.sub; // tu ID viene en el payload JWT
+    const id_usuario_b = id_amigo;
+
+    if (id_usuario_a === id_usuario_b) {
+      throw new BadRequestException('No podés eliminarte a vos mismo');
+    }
+
     return this.amistadService.deleteAmistad(id_usuario_a, id_usuario_b);
   }
 }
